Cambios double-decimal en entidades

+++ HorariosConsoleApp/Entities/DetallePagoEmpleado.cs
@@ -4,7 +4,7 @@
     {
         public int DetallePagoEmpleadoId { get; set; }
         public int PagoEmpleadoId { get; set; }
-        public double CantidadHoras { get; set; }
+        public decimal CantidadHoras { get; set; }
         public string TipoHora { get; set; }
         public decimal Porcentaje { get; set; }
     }
